% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fits_multiple.R,
%   R/compare_network_fits_multiple.R
\name{compare_fits_multiple}
\alias{compare_fits_multiple}
\title{compare_fits_multiple}
\usage{
compare_fits_multiple(formula_in, data = dat)

compare_fits_multiple(formula_in, data = dat)
}
\arguments{
\item{formula_in}{the input formula, has to be created with formula()}

\item{data}{the used dataframe for the survival analysis}

\item{formula_in}{the input formula, has to be created with formula()}

\item{data}{the used dataframe for the survival analysis}
}
\description{
This function fits multiple survival models for experience sampling data for further analysis
into a tidy dataframe

This function fits multiple survival models for experience sampling data for further analysis
into a tidy dataframe
}
\examples{
preprocess_exp_data
preprocess_exp_data
}
\keyword{preprocessing}
