% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_occurance.R
\name{desc_occurance}
\alias{desc_occurance}
\title{A function for the description of experience sampling data}
\usage{
desc_occurance(data)
}
\arguments{
\item{data}{the cleared dataframe}
}
\description{
This function gives discriptions of experience sampling data
}
\examples{
desc_occurance(dat1)

}
\keyword{descriptives}
\keyword{experience}
\keyword{sampling}
