% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_occurance_rolling.R
\name{desc_occurance_rolling}
\alias{desc_occurance_rolling}
\title{A function for the description of experience sampling data with lags}
\usage{
desc_occurance_rolling(data, lag = 2)
}
\arguments{
\item{data}{the cleared dataframe}

\item{lag}{the number of lags, default = 2}
}
\description{
This function gives discriptions of experience sampling data with lags
}
\examples{
desc_occurance_rolling(dat1, lag = 3)

}
\keyword{descriptives}
\keyword{experience}
\keyword{sampling}
