% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faceted_survival.R
\name{facet_survival_comp}
\alias{facet_survival_comp}
\title{facets of a fitted survival object}
\usage{
facet_survival_comp(fit, data, facet_vector, ncol = 2, title = NULL)
}
\arguments{
\item{fit}{the fitted object}

\item{data}{the dataframe (or tibble) used to fit the survfit object}

\item{facet_vector}{the variable(s) to facet by, could be 1+ , must be a string}

\item{title}{obtional title for the plot}

\item{number}{of columns, default = 2}
}
\description{
This function allows you facet a fitted survfit object. Generally, it is an modified version of
ggsurvplot_facet which uses geom_smooth for more comparison of the survival curve.
}
\examples{
facet_survival(fit_survival, dat, "beep", ncol = 3, title = "Faceted Survival Function")
}
\keyword{faceting}
