% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_survival_network.R
\name{facet_survival_network}
\alias{facet_survival_network}
\title{facet_survival_network}
\usage{
facet_survival_network(data)
}
\arguments{
\item{data}{the used dataframe for the survival analysis}

\item{formula_in}{the input formula, has to be created with formula()}
}
\description{
This function fits survival models for network data, facetted by a variable
into a tidy dataframe
}
\examples{
preprocess_exp_data


This function preprocesses experience sampling data for further analysis into a tidy dataframe
preprocess_exp_data
}
\keyword{preprocessing}
