% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{install_test}
\alias{install_test}
\title{A Test function}
\usage{
install_test(test = TRUE)
}
\arguments{
\item{test}{Did you install the package by purpose? Defaults to TRUE.}
}
\description{
This function allows you to test your package.
}
\examples{
install_test()
}
\keyword{test}
