% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occurance.R
\name{plot_occurance}
\alias{plot_occurance}
\title{A function for the plotting of occurances in experience sampling data}
\usage{
plot_occurance(data, title = "")
}
\arguments{
\item{data}{the cleared dataframe}

\item{title}{the title (optional)}
}
\description{
This function produces a plot of occurances of experience sampling data, facetted by ID
}
\examples{
plot_occurance(dat1)

}
\keyword{descriptives}
\keyword{experience}
\keyword{sampling}
