% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occurance_rolling.R
\name{plot_occurance_rolling}
\alias{plot_occurance_rolling}
\title{A function for the plotting of occurances in experience sampling data with lagged data}
\usage{
plot_occurance_rolling(data, title = "")
}
\arguments{
\item{data}{an object, produced with desc_occurance_rolling}

\item{title}{the title (optional)}
}
\description{
This function produces a plot of occurances of experience sampling data
}
\examples{
plot_occurance(dat1)

}
\keyword{descriptives}
\keyword{experience}
\keyword{sampling}
