% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_exp_data.R
\name{preprocess_exp_data}
\alias{preprocess_exp_data}
\title{preprocess experience sampling data}
\usage{
preprocess_exp_data(data)
}
\description{
This function preprocesses experience sampling data for further analysis into a tidy dataframe
}
\examples{
preprocess_exp_data
}
\keyword{preprocessing}
