% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampling.R
\name{sample_ex_data}
\alias{sample_ex_data}
\title{A generell sample function for the creation of Experience Sampling Data}
\usage{
sample_ex_data(n_participants = 100, n_days = 7)
}
\arguments{
\item{n_participants}{Number of Participants. Default = 100}

\item{n_days}{Number of Days. Default = 7}
}
\description{
This function allows you to test sample experience data.
}
\examples{
data_sample <-  sample_ex_data(50, 10)
}
\keyword{experience}
\keyword{sampling}
